finish exercise 4.22

+++ part4/bloglist/controllers/blogs.js
@@ -1,10 +1,16 @@
 const blogsRouter = require("express").Router();
-const { response } = require("express");
 const _ = require("lodash");
-const { request } = require("../app");
-
 const Blog = require("../models/blog");
-const { blogsInDb } = require("../tests/test_helper");
+const User = require("../models/user");
+const jwt = require("jsonwebtoken");
+
+const getTokenFrom = (request) => {
+  const authorization = request.get("authorization");
+  if (authorization && authorization.toLowerCase().startsWith("bearer ")) {
+    return authorization.substring(7);
+  }
+  return null;
+};
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({});
@@ -14,19 +20,50 @@ blogsRouter.get("/", async (request, response) => {
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
 
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+
+  // if (!body.userId){
+  //   return response.status(400).json({ error: "User ID missing" });
+  // }
+
+  // const userBlog = await User.findById(body.userId)
+
   const newBlog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user._id,
   });
   const savedBlog = await newBlog.save();
+
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   response.json(savedBlog);
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+
+  const blog =await Blog.findById(request.params.id);
+  // console.log(blog.user, user.id);
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(401).json({ error: "invalid token" });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
